// src/components/Flashcard/types.ts
export type FlipCallback = (() => void) | null;

export default interface FlashcardProps {
  /**
   * Unique identifier for the card
   */
  id?: string | number;
  /**
   * HTML string or JSX element to be displayed on the front of the card
   */
  frontHTML: string | JSX.Element;
  /**
   * CSS styles to be applied to the front side of the card
   */
  frontCardStyle: React.CSSProperties | string;
  /**
   *  CSS styles to be applied to the content of the front side of the card
   */
  frontContentStyle: React.CSSProperties | string;
  /**
   * HTML string or JSX element to be displayed on the back of the card
   */
  backHTML: string | JSX.Element;
  /**
   * CSS styles to be applied to the back side of the card
   */
  backCardStyle: React.CSSProperties | string;
  /**
   * CSS styles to be applied to the content of the back side of the card
   */
  backContentStyle: React.CSSProperties | string;
  /**
   * CSS class to be applied to the wrapper div
   */
  className?: string;
  /**
   * CSS height of the wrapper div
   */
  height?: string;
  /**
   * CSS border-radius of the wrapper div
   */
  borderRadius?: string;
  /**
   * CSS width of the wrapper div
   */
  width?: string;
  /**
   * CSS styles to be applied to the wrapper div
   */
  style?: React.CSSProperties;
  /**
   *  Callback function to be called when the card is flipped
   */
  onCardFlip?: (state: boolean) => void;
  /**
   * Ref object that exposes the flipCard function to manually control card flipping
   * @example
   * const flipRef = useRef<() => void>(null);
   * // Later: flipRef.current?.()
   */
  manualFlipRef?: React.MutableRefObject<FlipCallback>;
}


export interface FlashcardRef {
  onManualFlip: () => void;
  isFlipped: boolean;
}